Add deidentify_videos.py
deidentify_videos.py is a python script that will take a directory of
videos, strip the excess metadata with ffmpeg, optionally randomize the
filename (as either a uuid or a successively incrementing number), then
place the videos into a new directory. The name translation is stored in
a CSV log file.