Add pp_dd.R
R script that prints to standard output markdown to document R
dataframes/tibbles. Requires an RDS file that is either 1. A dataframe
with columns (name, df), with name, a character vector, and df, a list
column holding dataframes/tibbles or 2. a named list of dataframes.

It will output a markdown heading with the table name followed by
a markdown table with columns \"name\", \"class\" (eg., integer,
character, etc), and \"description\" (blank column for you to fill in).