Stop making tempdir if no videos in the dir
Move the creation of a temporary directory to occur after reading a list
of the videos in the directory. That way, if there are no videos, a
temporary directory will not be created.

I still am not going to have it delete the tempdir if it errors after
this point, just so that the person running can see the files that were
made and use them to help with debugging.